% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCutilities-package.R
\docType{package}
\name{ABCutilities-package}
\alias{ABCutilities}
\alias{ABCutilities-package}
\title{ABCutilities: Generally useful functions used by multiple ABC packages}
\description{
Functions for error checking and other useful things, including numerous helper functions for our shiny apps.
}
\author{
\strong{Maintainer}: Applied Bioinformatics Core Weill Cornell Medicine \email{abc@med.cornell.edu}

Other contributors:
\itemize{
  \item Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [contributor]
}

}
\keyword{internal}
