% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{color_grid_facet_bg}
\alias{color_grid_facet_bg}
\title{Add user-specified colors to facet grid strips}
\usage{
color_grid_facet_bg(in_plot, facet_cols)
}
\arguments{
\item{in_plot}{ggplot2 object generated with facet_grid (see examples).}

\item{facet_cols}{vector of color names, e.g. c("red","green","blue").
Its length should correspond to the factor that was used to assign the
\code{facet_grid} with.}
}
\value{
This function will print the newly colored plot to the
graphics device.
}
\description{
This function is useful to assign specific colors to the
background of the facet_grid labels.
}
\details{
See
\url{https://stackoverflow.com/questions/53455092/r-ggplot2-change-colour-of-font-and-background-in-facet-strip}
for details.
This routine will only work with facet_grid (not facet_wrap) and
\code{in_plot} should \emph{not} have the background color stripped from it via
the \code{theme} setting
}
\examples{
\dontrun{
# generate base plot
Pp <- ggplot(tmp2, aes(x = HPCA.main, y = N)) +
  geom_bar(stat = "identity", aes(fill = condition)) +
  scale_fill_manual(values = condition_cols) +
  facet_grid(cluster ~ .)

color_grid_facet_bg(Pp, facet_cols = cluster_cols)
}

}
