% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{fx.get_palette_ABC}
\alias{fx.get_palette_ABC}
\title{Color palettes}
\usage{
fx.get_palette_ABC(palette_name)
}
\value{
vector of color names
}
\description{
This function simply supplies pre-defined color palettes.
}
\details{
Based on \code{scater}'s defaults, but with significant changes to the
standard colors that were bing used
}
\seealso{
\code{fx.resolve_plot_colors}
}
