% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{fx.parse_column_names}
\alias{fx.parse_column_names}
\title{Parse column names}
\usage{
fx.parse_column_names(cn)
}
\arguments{
\item{cn}{column name}
}
\value{
column name enclosed in single quotes
}
\description{
Parse column names
}
\details{
There are numerous ways to access the values
stored in the columns of a \code{data.frame}. The one
using a dollar sign (e.g. \code{df$this_column}) is
vulnerable to special characters, such as "-", spaces etc.
Since the default method of \code{ggplot2} is to use the
dollar sign approach, we need to ensure that special
characters, which are often part of gene names, are taken
care of. This function simply wraps those names into
single quotes.
}
