% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{fx.resolve_plot_colors}
\alias{fx.resolve_plot_colors}
\title{Get nice plotting color schemes for very general color variables}
\usage{
fx.resolve_plot_colors(plot_out, colour_by, colour_by_name, fill = FALSE)
}
\arguments{
\item{plot_out}{ggplot2 object}

\item{colour_by}{vector of values that determine the coloring of \code{plot_out}}

\item{colour_by_name}{string indicating the title/name for \code{colour_by},
e.g. the name of a gene}

\item{fill}{Boolean, default: 'FALSE'}
}
\value{
\code{ggplot2} object with adjusted coloring scheme
}
\description{
Wrapper around \code{fx.get_palette_ABC} that checks the numbers
and types of coloring variables that are being used and tries to return a
sensible color scheme. The colors are pre-defined in \code{fx.get_palette_ABC}.
}
\details{
This function is based on a very similar function in the scater package.
}
\seealso{
\code{fx.get_palette_ABC}
}
