% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{fx.resolve_plot_shapes}
\alias{fx.resolve_plot_shapes}
\title{Supply more than 6 somewhat meaningful shapes}
\usage{
fx.resolve_plot_shapes(plot_in, shape_by)
}
\arguments{
\item{plot_in}{ggplot2 object where shape is one of the aesthetics}

\item{shape_by}{the factors to which shape has been mapped}
}
\value{
ggplot2 object with the shape scale set manually
}
\description{
Supply more than 6 somewhat meaningful shapes
}
\details{
The ggplot2 default setting will complain about having more than 6
different factors assigned to shape. This function assigns 12 shapes (if
there's more than 6 factors in \code{shape_by}), trying to find a good order
that will still allow for differentiations.
}
