% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{get_legend}
\alias{get_legend}
\title{Retrieve the legend of a plot}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{A \code{ggplot} object that was created with a legend}
}
\value{
A \code{gtable} object holding just the legend
}
\description{
This function extracts just the legend from a \code{ggplot}.
}
\details{
From
\url{https://stackoverflow.com/questions/11883844/inserting-a-table-under-the-legend-in-a-ggplot2-histogram}
}
\examples{
p1 <- ggplot2::ggplot(
  datasets::mtcars,
  ggplot2::aes(mpg, disp)
) +
  ggplot2::geom_line()
plot.mpg <- ggplot2::ggplot(
  ggplot2::mpg,
  ggplot2::aes(x = cty, y = hwy, colour = factor(cyl))
) +
  ggplot2::geom_point(size = 2.5)
legend <- get_legend(plot.mpg)
# plot the plot
plot.mpg <- plot.mpg + ggplot2::theme(legend.position = "none")
# plot the legend
grid::grid.draw(legend)

}
