% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gorilla2REVIGO.R
\name{GOenrichment_GOrilla_REVIGO}
\alias{GOenrichment_GOrilla_REVIGO}
\title{Submit genes to GOrilla and retrieve REVIGO data}
\usage{
GOenrichment_GOrilla_REVIGO(organism = NULL, gorilla_runmode = "hg",
  targets = NULL, background = NULL, outdir = NULL,
  analysis_name = NULL, perl_path = "/opt/local/bin/perl",
  python_path = "/usr/local/bin/python")
}
\arguments{
\item{organism}{either one of "ARABIDOPSIS_THALIANA", "SACCHAROMYCES_CEREVISIAE",
"CAENORHABDITIS_ELEGANS", "DROSOPHILA_MELANOGASTER","DANIO_RERIO","HOMO_SAPIENS",
"MUS_MUSCULUS", "RATTUS_NORVEGICUS"}

\item{targets}{vector of gene symbols for which the GO terms will be retrieved}

\item{background}{vector of gene symbols for which the GO terms will be retrieved and compared to the targets}

\item{outdir}{path to directory where results should be stored}

\item{analysis_name}{a label for the analysis, e.g. "DB_condition", or "cluster1"}

\item{perl_path}{Define which perl installation should be used to run the GORILLA
script. Set to "/opt/local/bin/perl".}

\item{python_path}{Define which python installation should be used to run the 
REVIGO script. Default: "/usr/local/bin/python"}

\item{runmode:}{either 'mhg' (one ranked target list) or 'hg' (target and background list)}
}
\value{
A list of 3, one for PROCESS, FUNCTION and COMPONENT. Each list contains
the information that are needed to draw a REVIGO treemap: the GO terms, their 
grouping (e.g. "circadian rhythm") and several REVIGO stats. If a specific GO
category did not yield any enriched results via GOrilla, that category will be
missing from the resulting list here.
}
\description{
Submit genes to GOrilla and retrieve REVIGO data
}
\details{
The GORILLA and REVIGO scripts are part of the installation of
ABCutilities (\code{system.file("extdata", "gorilla_submission.pl", package = "ABCutilities")} 
and \code{system.file("extdata", "download_revigo_improved.py", package = "ABCutilities")}).
}
\examples{
\dontrun{
rev_res_for_treemap <- GOenrichment_GOrilla_REVIGO(organism = "HOMO_SAPIENS", 
                            targets = "sandbox/forGOrilla_gD.txt",
                            background = "sandbox/forGOrilla_backgroundPx.txt", 
                            outdir = "sandbox/test3/", 
                            gorilla_runmode = "hg", analysis_name = "blubb")
}

## to generate the REVIGO maps
REVIGO_treemap(revigo.data = rev_res_for_treemap$test_PROCESS, col_palette = "Set3",title = "testing")

}
