% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gorilla2REVIGO.R
\name{GOrilla2Revigo}
\alias{GOrilla2Revigo}
\title{Automated submission of GOrilla results to REVIGO}
\usage{
GOrilla2Revigo(
  GOrilla_link,
  analysis_name = NULL,
  out_path = NULL,
  python_path = "/usr/local/bin/python"
)
}
\arguments{
\item{GOrilla_link}{either the full link to a GOrilla result,
e.g. "http://cbl-gorilla.cs.technion.ac.il/GOrilla/oxso2ua3" or
just the unique identifier ("oxso2ua3").
Can also be the \strong{path} to an already downloaded html file.}

\item{analysis_name}{a label for the analysis,
e.g. "DB_condition", or "cluster1".}

\item{out_path}{path to where the (temporary) results are to be stored.}

\item{python_path}{path to Python executable.}
}
\value{
A list of 3, one for PROCESS, FUNCTION and COMPONENT.
Each list contains the information that are needed
to draw a REVIGO treemap: the GO terms, their
grouping (e.g. "circadian rhythm") and several REVIGO stats.
If no GO terms were enriched, this function will return \code{NULL}.
}
\description{
This function takes as input the link to an individual GOrilla
result, which it uses to download the results, which are then submitted to
REVIGO.
}
\details{
This functions makes all kinds of assumptions that may render it not
as robust as other functions. It needs python and sed; make sure the Python
pass is the correct one. If NULL, it will use the output of
\code{system("which python")}, which may not be approriate.
Default: "/usr/local/bin/python".

It also assumes that GOrilla has been executed for all GO term classes,
i.e. PROCESS, FUNCTION, and COMPONENT.
}
\examples{
\dontrun{
test <- GOrilla2Revigo("gwlk3omz", "test")
# List of 3
# test_PROCESS  : chr [1:158, 1:7] "GO:0007623" "GO:0008152" "GO:0009987" ...
# test_FUNCTION : chr [1:63, 1:7] "GO:0001228" "GO:0003700" "GO:0001077" ...
# test_COMPONENT: chr [1:63, 1:7] "GO:0031974" "GO:0032991" "GO:0043226" ...
test2 <- GOrilla2Revigo("GOrilla_results_COMPONENT_test.txt")
# List of 3
# PROCESS  : chr [1:63, 1:7] "GO:0031974" "GO:0032991" "GO:0043226" ...
# FUNCTION : chr [1:63, 1:7] "GO:0031974" "GO:0032991" "GO:0043226" ...
# COMPONENT: chr [1:63, 1:7] "GO:0031974" "GO:0032991" "GO:0043226" ...

## to generate the REVIGO maps
REVIGO_treemap(
  revigo.data = test$test_PROCESS, col_palette = "Set3", title = "testing"
)
}

}
\seealso{
\link{REVIGO_treemap}, \link{submit_to_gorilla}
}
