% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{jgc}
\alias{jgc}
\title{Java garbage collection function}
\usage{
jgc()
}
\description{
To avoid crashing of \code{xlsx::write.xlsx()}.
Ideally, just make use of \code{openxlsx} instead of \code{xlsx},
this removes the Java dependency all together.
}
