% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplotlist.R
\name{MultiPlotList}
\alias{MultiPlotList}
\title{Plot multiple ggplot2 objects on one page}
\usage{
MultiPlotList(plots, cols = 1, layout = NULL, title = NULL)
}
\arguments{
\item{plots}{list of ggplot2 objects}

\item{cols}{number of columns that the resulting plot should have}

\item{layout}{matrix with ncol and nrow, if left to NULL it will be deter-
mined automatically}

\item{title}{supply a title that will be printed across all columns (default:NULL)}
}
\description{
Plot multiple ggplot2 objects on one page
}
\examples{
\dontrun{
label_list <- list(sample = "sample", replicate = "technical_replicate")
## make a list of ggplot2 objects
pl <- lapply(seq_along(label_list),  function(i){
       P <-  draw_PCA_results(sceset.filt,
                      tsne_mg$all$V2, tsne_mg$all$V3,
                      exprs_values = "log10magic",
                      colour_by = names(label_list[i]))
       P <- P + ggtitle(label_list[[i]]) + theme_bw(base_size = 16)
     })
 ## plot them all on the same page
 MultiPlotList(pl, cols = 2, title = "tsne log10(magic)")
}
}
