% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{n2n}
\alias{n2n}
\title{Helper function for shiny scripts}
\usage{
n2n(x)
}
\description{
Sometimes, it is more convenient to allow the user to select 
"none" instead of setting a variable to NULL. Within the scripts, however,
it is often easier to handle instances of NULL than of the string "none".
This function is supposed to help with that as it returns NULL if x is either
"none" or an empty string.

On-the-fly NULLing of "none" or "" entries.
}
