% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_breaks.R
\name{quantile_breaks}
\alias{quantile_breaks}
\title{Calculating quantile breaks, e.g. for usage with pheatmap settings}
\usage{
quantile_breaks(xs, n = 10)
}
\arguments{
\item{xs}{matrix or vector of numbers}

\item{n}{number of breaks}
}
\value{
named vector of breaks where the names indicate the quantiles
and the numbers indicate the actual values where the breaks are to
be drawn
}
\description{
Function from Kamil Slowikowski (\url{https://slowkow.com/notes/heatmap-tutorial/}).
}
\examples{
hm_mat <- matrix(rnorm(200), 20, 10)
hm_mat[c(100, 101)] <- max(hm_mat) + 1000

mat_breaks <- quantile_breaks(hm_mat, n = 10)
pheatmap::pheatmap(hm_mat,
  breaks = mat_breaks,
  color = grDevices::colorRampPalette(
    rev(RColorBrewer::brewer.pal(n = 7, name = "RdYlBu"))
  )(length(mat_breaks) - 1)
)

}
