% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{REVIGO_treemap}
\alias{REVIGO_treemap}
\title{Makes a treemap based on data exported by REVIGO}
\usage{
REVIGO_treemap(revigo.data, col_palette = "Paired",
  title = "REVIGO Gene Ontology treemap", ...)
}
\arguments{
\item{revigo.data}{the result of \code{\link{GOrilla2REVIGO}}, i.e. a data.frame
with GO terms and the respective REVIGO stats}

\item{col_palette}{which color palette to use, e.g. "Paired" (default),
"Spectral", "Accent", "Set1", "Set2", "-RdGy", or \code{terrain.colors(10)}}

\item{title}{title to be printed on top of the plot}

\item{...}{additional parameters for \code{treemap::treemap}}
}
\description{
Makes a treemap based on data exported by REVIGO
}
\details{
This function is based on the original R scripts that are being
generated by REVIGO \url{(http://revigo.irb.hr/revigo.jsp)}.
}
\examples{
\dontrun{
revigo.data <- rbind(c("GO:0005576","extracellular region",4.572,5.7113,0.960,0.000,"extracellular region"),
                    c("GO:0005623","cell",64.133,17.3884,0.985,0.000,"cell"),
                    c("GO:0005739","mitochondrion",3.808,17.2418,0.442,0.000,"mitochondrion"),
                    c("GO:0045251","electron transfer flavoprotein complex",0.000,4.1740,0.845,0.183,"mitochondrion"),
                    c("GO:0005737","cytoplasm",38.159,31.1624,0.773,0.221,"mitochondrion"),
                    c("GO:0017133","mitochondrial electron transfer flavoprotein complex",0.000,4.1740,0.684,0.234,"mitochondrion"))

REVIGO_treemap(revigo.data, title="REVIGO treemap: cellular components")
}

}
\seealso{
\code{\link{GOrilla2Revigo}}
}
