% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterms_GOseqToRevigo.R
\name{txtFile2REVIGO}
\alias{txtFile2REVIGO}
\title{Text file submission to REVIGO}
\usage{
txtFile2REVIGO(infile, analysis_name = NULL, out_path = NULL,
  python_path = "/usr/local/bin/python")
}
\arguments{
\item{infile}{path to a txt file with GO terms and p-values}

\item{analysis_name}{a label for the analysis, e.g. "DB_condition", or "cluster1"}

\item{out_path}{path to where the (temporary) results are to be stored}

\item{python_path}{Define which python installation should be used to run the 
REVIGO script. Default: "/usr/local/bin/python"}
}
\value{

}
\description{
This function expects a txt file with two columns (GO term, p-value)
that will be submitted to REVIGO. The result will be a data.frame that can be
used to generate a treemap. If you want to start from a list of genes, consider
\code{\link{GOenrichment_GOrilla_REVIGO}}.
}
\details{
The REVIGO API will be accessed with the help of a python script that
is part of the ABCutilities package. This function needs python and sed;
make sure the Python path is the correct one. If NULL, it will use the output
of \code{system("which python")}, which may not be approriate. Default: "/usr/local/bin/python".
}
\examples{
\dontrun{
test <- txtFile2REVIGO("data/rna-seq_deseq2_WT_vs_KO_0.05_revigo_in.txt", analysis_name = "testing")
#' head(test)
#[,1]         [,2]                                           [,3]    [,4]     [,5]    [,6]    [,7]                      
#[1,] "GO:0001775" "cell activation"                              "0.171" "6.587"  "0.781" "0"     "cell activation" 
#[2,] "GO:0002250" "adaptive immune response"                     "0.053" "7.6594" "0.418" "0"     "adaptive immune response"
#[3,] "GO:0007165" "signal transduction"                          "6.621" "7.012"  "0.562" "0.675" "adaptive immune response"
#[4,] "GO:0007186" "G-protein coupled receptor signaling pathway" "0.882" "5.676"  "0.613" "0.301" "adaptive immune response"
#[5,] "GO:0006935" "chemotaxis"                                   "0.475" "4.3015" "0.723" "0.366" "adaptive immune response"
#[6,] "GO:0002376" "immune system process"                        "0.6"   "4.5066" "0.915" "0"     "immune system process"

## to generat the REVIGO maps
REVIGO_treemap(revigo.data = test, col_palette = "Set3",title = "testing")

}

}
\seealso{
\code{\link{REVIGO_treemap}}, \code{\link{GOrilla2Revigo}}
}
