% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gorilla2REVIGO.R
\name{fx.retrieve_GOrilla_results}
\alias{fx.retrieve_GOrilla_results}
\title{Helper function for specifying where to get the GOrilla results from}
\usage{
fx.retrieve_GOrilla_results(GOrilla_link, TYPE, out_path, analysis_name)
}
\description{
This function handles the different types of entries that can be
supplied to \code{GOrilla_link}.
}
