% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{get_legend}
\alias{get_legend}
\title{Retrieve the legend of a plot}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{A \code{ggplot} that was created with a legend}
}
\value{
A \code{gtable} object holding just the legend
}
\description{
This function extracts just the legend from a \code{ggplot}.
}
\details{
From \url{https://stackoverflow.com/questions/11883844/inserting-a-table-under-the-legend-in-a-ggplot2-histogram}
}
\examples{
p1 <- ggplot(mtcars, aes(mpg, disp)) + geom_line()
plot.mpg <- ggplot(mpg, aes(x = cty, y = hwy, colour = factor(cyl))) + geom_point(size=2.5)
legend <- get_legend(plot.mpg)
# plot the plot
plot.mpg <- plot.mpg + theme(legend.position='none')
# plot the legend
grid.draw(legend)
}
