% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterms_GOseqToRevigo.R
\name{txtFile2REVIGO}
\alias{txtFile2REVIGO}
\title{Text file submission to REVIGO}
\usage{
txtFile2REVIGO(
  infile,
  analysis_name = NULL,
  out_path = NULL,
  python_path = "/usr/local/bin/python"
)
}
\arguments{
\item{infile}{path to a txt file with GO terms and p-values}

\item{analysis_name}{a label for the analysis,
e.g. "DB_condition", or "cluster1"}

\item{out_path}{path to where the (temporary) results are to be stored}

\item{python_path}{Define which python installation should be used to run the
REVIGO script. Default: "/usr/local/bin/python"}
}
\value{
revigo data, whatever it may be.
}
\description{
This function expects a txt file with two columns
(GO term, p-value) that will be submitted to REVIGO.
The result will be a data.frame that can be used to generate a treemap.
If you want to start from a list of genes,
consider \code{\link[=GOenrichment_GOrilla_REVIGO]{GOenrichment_GOrilla_REVIGO()}}.
}
\details{
The REVIGO API will be accessed with the help of a python script
that is part of the ABCutilities package.
This function needs python and sed;
Make sure the Python path is the correct one.
If NULL, it will use the output of \code{system("which python")},
which may not be approriate. Default: "/usr/local/bin/python".
}
\examples{
\dontrun{
test <- txtFile2REVIGO(
  "data/rna-seq_deseq2_WT_vs_KO_0.05_revigo_in.txt",
  analysis_name = "testing"
)
head(test)
## to generat the REVIGO maps
REVIGO_treemap(
  revigo.data = test,
  col_palette = "Set3",
  title = "testing"
)
}

}
\seealso{
\code{\link{REVIGO_treemap}}, \code{\link{GOrilla2Revigo}}
}
